Renames settings->build_helpers
