+++ frigg_worker/api.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 import logging
-
 import requests
 
 logger = logging.getLogger(__name__)
