+++ frigg_worker/jobs.py
@@ -8,7 +8,7 @@ import requests
 from frigg.config import config, sentry
 from frigg.helpers import cached_property
 from frigg_coverage import parse_coverage
-from frigg_worker.settings import build_settings
+from frigg_worker.build_helpers import build_settings
 
 from . import api
